% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P2C2M_GMYC_pboot.R
\name{P2C2M_GMYC.pboot}
\alias{P2C2M_GMYC.pboot}
\title{FUNCTION_TITLE}
\usage{
P2C2M_GMYC.pboot(
  tree.input,
  tree.format,
  seq,
  nsim = NULL,
  nboot = NULL,
  perc.treshold = NULL
)
}
\arguments{
\item{tree.input}{PARAM_DESCRIPTION}

\item{tree.format}{PARAM_DESCRIPTION}

\item{seq}{PARAM_DESCRIPTION}

\item{nsim}{PARAM_DESCRIPTION, Default: NULL}

\item{nboot}{PARAM_DESCRIPTION, Default: NULL}

\item{perc.treshold}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[devtools]{remote-reexports}}
}
