% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P2C2M_GMYC_pboot.R
\name{P2C2M_GMYC.pboot}
\alias{P2C2M_GMYC.pboot}
\title{Run P2C2M.GMYC - parametric bootstrap}
\usage{
  P2C2M_GMYC.pboot(tree.input, tree.format, seq, nsim = NULL,
     nboot = NULL, perc.treshold = NULL)
}
\arguments{
  \item{tree.input}{A file name which contains an ultrametric phylogenetic tree.}

  \item{tree.format}{A character string specifying the format of the phylogenetic tree. Two choices are possible: "nexus" or "newick".}

  \item{seq}{A file name which contains a sequential FASTA alignment}

  \item{nsim}{Number of simulated trees, Default: 100 trees}

  \item{nboot}{The size of the null distribution, Default: 100}

  \item{perc.treshold}{Percentage of the number of species recovered by the empirical dataset used to set the threshold to calculate the p-value, Default: 0.1 (10\%)}
}
\value{
  A list containing the p-value, the threshold used to calculate the p-value, and the null distribution.
}
\description{
  Identifying model violations under the Generalized Mixed Yule Coalescent (GMYC) model using parametric bootstrap.
}
\details{
  This function take a ultrametric phylogenetic tree and a sequential FASTA alignment and test for model violation of the GMYC model using a parametric bootstrap framework.
}
\examples{
## Lygodactylus dataset

P2C2M_GMYC.pboot(tree.input=system.file("extdata", "Lygodactylus.tre", package="P2C2M.GMYC"),
                 tree.format="nexus",
                 seq=system.file("extdata", "Lygodactylus.fas", package="P2C2M.GMYC"))
  }
}
