% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P2C2M_GMYC_pboot.R
\name{P2C2M_GMYC.pboot}
\alias{P2C2M_GMYC.pboot}
\title{Parametric bootstrap}
\usage{
P2C2M_GMYC.pboot(tree.input, tree.format, seq, nsim = NULL, nboot = NULL, perc.treshold = NULL)
}
\arguments{
\item{tree.input}{a file name which contains an ultrametric phylogenetic tree.}

\item{tree.format}{a character string specifying the format of the phylogenetic tree. Two choices are possible: "nexus" or "newick"}

\item{seq}{a file name which contains a sequential FASTA alignment}

\item{nsim}{Number of simulated trees, Default: 100 trees}

\item{nboot}{The size of the null distribution, Default: 100}

\item{perc.treshold}{Percentage of the number of species recovered by the empirical dataset used to set the threshold to calculate the p-value, Default: 0.1}
}
\value{
A list containing the p-value, threshold used to calculate the p-value, and the null distribution.
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[devtools]{remote-reexports}}
}
